chore: move to storybook recipes
