--- packages/blade/src/components/NativeAria/index.ts
@@ -1,2 +1,0 @@
-export { default } from './NativeAria.web';
-export * from './NativeAria.web';
