+++ packages/blade/src/storybook-recipes/AccessibilityLayer/NativeAria.web.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import mapA11yProps from '../../utils/mapProps';
-import { CloseIcon, EyeIcon } from '../Icons';
+import { CloseIcon, EyeIcon } from '../../components/Icons';
 import { getCheckboxA11yProps } from './getCheckboxA11yProps';
 
 const Checkbox: React.FC<{ label: string; checked?: boolean; disabled?: boolean }> = ({
